health_service/requirements.txt
Flask>=2.0
psycopg2-binary>=2.9
requests>=2.28
azure-identity>=1.12.0
